updates to people
We may need to make a couple of changes to the web, it all possible. From the EAC (under people), Dr. Juliane needs to be removed. From the internal advisory committee, Dr. Kaushal, Dr. Reed and Dr. Kolls need to be removed. We may need to add Dr. Tammy Dugas and Dr. Udeni Balasuriya.

Thanks,
Jey